% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asmcall.R
\name{assemble}
\alias{assemble}
\title{Asemble a string of assembly code}
\usage{
assemble(asm, flavor = "nasm")
}
\arguments{
\item{asm}{A scalar string of assembly code}
}
\value{
A handle to the linked assembly code
}
\description{
Currently only NASM is supported.
}
\details{
Supported sections;
\itemize{
\item .data
\item .rodata
\item .text
\item .rela.text
\item .rel.text
}

Runtime linked APIs such as R APIs can be declared as \code{extern}.
}
