% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asmcall.R
\name{.Asm}
\alias{.Asm}
\title{Call an assembly function in the .C calling convention}
\usage{
.Asm(box, name, ...)
}
\arguments{
\item{box}{A handle to the linked assembly code}

\item{name}{The name of the function to call}

\item{...}{Arguments to pass to the assembly function}
}
\value{
The return value of the assembly function
}
\description{
The underlying code are expected to accept SEXP's and return one SEXP.
}
