% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasm.R
\docType{package}
\name{rasm}
\alias{rasm-package}
\alias{rasm}
\title{Rasm: Inline Assembler for R}
\description{
This package provides a way to write inline assembly code in R.
}
\details{
It is mostly a proof of concept and should not be used in production.

However it has the following "advantages" to traditional ways to write R extensions (mostly joking):
\enumerate{
\item CRAN friendly: CRAN disallow "binary executable code" in packages, but
does not disallow unlinked assembly code that links itself at runtime.
\item No need to get a working compiler with R headers,
Just write the name of the function you need and this package will call it for you!
\item Hot reloading: Replace your native code without restarting R! Just delete the old page and create a new one.
\item Utmost compatibility: If R runs on this machine, your code will 100\% run too!
\item Unbeatable control: You can do anything you want, even if it's a bad idea.
}
}
\author{
\strong{Maintainer}: Anne Fu \email{anne.fu@utdallas.edu} (\href{https://orcid.org/0000-0002-9025-6071}{ORCID})

}
